Refine docs.
